+++ apps/workshop/.storybook/preview.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { UiProvider } from "@companydotcom/ui";
+import { UiProvider } from "@vastly/ui";
 
 export const parameters = {
   actions: { argTypesRegex: "^on[A-Z].*" },