+++ apps/workshop/pages/_app.tsx
@@ -1,5 +1,5 @@
 import type { AppProps } from "next/app";
-import { UiProvider } from "@companydotcom/ui";
+import { UiProvider } from "@vastly/ui";
 
 export default function MyApp({ Component, pageProps }: AppProps) {
   return (