+++ packages/forms/src/components/step-form/step-form.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 
 import { FieldValues } from "react-hook-form";
 import { chakra, Button, ButtonProps, HTMLChakraProps, ThemingProps } from "@chakra-ui/react";
-import { callAllHandlers, runIfFn, cx, isDev } from "@companydotcom/utils";
+import { callAllHandlers, runIfFn, cx, isDev } from "@vastly/utils";
 import {
   StepperProvider,
   NumberStepperSteps,
@@ -10,7 +10,7 @@ import {
   NumberStepperStep,
   useStepperContext,
   NumberStepperContainer,
-} from "@companydotcom/ui";
+} from "@vastly/ui";
 import { Form } from "../form";
 import { SubmitButton } from "../submit-button";