+++ packages/ui/src/components/pagination/pagination.tsx
@@ -14,7 +14,7 @@ import {
   ComponentWithAs,
   BoxProps,
 } from "@chakra-ui/react";
-import { __DEV__ } from "@companydotcom/utils";
+import { __DEV__ } from "@vastly/utils";
 import { usePagination, PageType } from "./use-pagination";
 import {
   PaginationProvider,