+++ apps/workshop/CHANGELOG.md
@@ -5,8 +5,8 @@
 ### Patch Changes
 
 - Updated dependencies [1ccfdc7]
-  - @companydotcom/forms@0.1.0
-  - @companydotcom/ui@0.1.0
+  - @vastly/forms@0.1.0
+  - @vastly/ui@0.1.0
 
 ## 0.0.2