+++ packages/cli/CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### Patch Changes
 
-- @companydotcom/utils@0.0.1
+- @vastly/utils@0.0.1
 
 ## 0.1.0