+++ packages/client/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @companydotcom/client
+# @vastly/client
 
 ## 0.1.0