+++ packages/test-utils/CHANGELOG.md
@@ -1,8 +1,8 @@
-# @companydotcom/test-utils
+# @vastly/test-utils
 
 ## 0.0.2
 
 ### Patch Changes
 
 - Updated dependencies [1ccfdc7]
-  - @companydotcom/ui@0.1.0
+  - @vastly/ui@0.1.0